--- apps/dotcms-ui/src/app/test/dot-timezone.mock.ts
@@ -1,19 +1,0 @@
-import { DotTimeZone } from '@services/dot-timezone/dot-timezone.service';
-
-export const mockDotTimeZones: DotTimeZone[] = [
-    {
-        id: 'America/Venezuela',
-        label: 'Venezuela',
-        offset: '240'
-    },
-    {
-        id: 'America/Costa Rica',
-        label: 'Costa Rica',
-        offset: '360'
-    },
-    {
-        id: 'America/Panama',
-        label: 'Panama',
-        offset: '300'
-    }
-];
