+++ apps/dotcms-ui/src/app/view/components/_common/forms/dot-push-publish-form/dot-push-publish-form.module.ts
@@ -11,7 +11,7 @@ import { DotParseHtmlService } from '@services/dot-parse-html/dot-parse-html.ser
 import { CalendarModule } from 'primeng/calendar';
 import { DropdownModule } from 'primeng/dropdown';
 import { SelectButtonModule } from 'primeng/selectbutton';
-import { DotTimeZoneService } from '@services/dot-timezone/dot-timezone.service';
+import { DotcmsConfigService } from '@dotcms/dotcms-js';
 
 @NgModule({
     declarations: [DotPushPublishFormComponent],
@@ -28,6 +28,6 @@ import { DotTimeZoneService } from '@services/dot-timezone/dot-timezone.service'
         SelectButtonModule,
         DotPipesModule
     ],
-    providers: [PushPublishService, DotParseHtmlService, DotTimeZoneService]
+    providers: [PushPublishService, DotParseHtmlService, DotcmsConfigService]
 })
 export class DotPushPublishFormModule {}
